+++ app/src/main/res/xml/single_blind_preferences.xml
@@ -2,7 +2,7 @@
 <PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".ui.blind.SingleBlindSettingsPreferencesFragment">
+    tools:context=".ui.blinds.SingleBlindSettingsPreferencesFragment">
 
     <EditTextPreference
         android:key="ip"
